Added PlayersList Component
