+++ src/pages/Dashboard/components/PlayerManagement/components/PlayersList/index.ts
@@ -1,0 +1,1 @@
+export {PlayersList} from './PlayersList';
